Issue #101: SocketTimeout causes rebuilding a PR multiple times.
Cascading IOException to fix issue #101
If a SocketTimeoutException is thrown while reading any of the
comment responses, it will be consumed and an empty list will be
returned for the comments. The empty list then triggers the build
to run again because the plugin thinks no previous attempts were made.